% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_allen_relations.r
\docType{data}
\name{allen_relations}
\alias{allen_relations}
\alias{stilf_relation_before}
\title{Allens Intervals}
\usage{
stilf_relation_before(first_interval , second_interval)
}
\arguments{
\item{first_interval}{stilf_interval. An interval between two dates.}

\item{second_interval}{stilf_interval. An interval between two dates.}
}
\value{
Logical value if interval are TRUE or FALSE
}
\description{
Provide a set of Allen's interval relations to classified time series data. And return a logical value if an interval is TRUE or FALSE
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# create some examples of intervals
time1 <- stilf_interval("2011-09-01","2011-10-01")
time2 <- stilf_interval("2011-09-15","2011-11-01")
time3 <- stilf_interval("2011-10-01","2011-11-01")
time4 <- stilf_interval("2011-10-01","2011-11-01")
time5 <- stilf_interval("2011-08-01","2011-09-01")
time6 <- stilf_interval("2011-08-15","2011-09-15")
time7 <- stilf_interval("2011-08-01","2011-09-15")
time8 <- stilf_interval("2011-08-15","2011-08-29")

# Two interval I and J, with:
# I <- ("2011-08-15","2011-08-29")
# J <- ("2011-09-01","2011-10-01")
# end_I equals to the final value of the interval, I == "2011-08-29"
# start_I equals to the begin value of the interval, I == "2011-08-15"
# end_J equals to the final value of the interval, J == "2011-10-01"
# start_J equals to the begin value of the interval, J == "2011-09-01"

# Apply a relation 'before' (end_I < start_J)
stilf_relation_before(time1,time4)

# Apply a relation 'after' (start_I > end_J)
stilf_relation_after(time4,time1)

# Apply a relation 'meets' (end_I == start_J)
stilf_relation_meets(time1,time3)

# Apply a relation 'met by' (end_J == start_I)
stilf_relation_met_by(time1,time5)

# Apply a relation 'overlaps' (start_I < start_J) & (end_I > start_J) & (end_I < end_J)
stilf_relation_overlaps(time1,time2)

# Apply a relation 'overlapped by' (end_I > start_J) & (start_I < end_J) & (end_I > end_J)
stilf_relation_overlapped_by(time1,time6)

# Apply a relation 'starts' (start_I == start_J) & (end_I < end_J)
stilf_relation_starts(time5,time7)

# Apply a relation 'started by' (start_I == start_J) & (end_I > end_J)
stilf_relation_started_by(time7,time5)

# Apply a relation 'during' (start_I > start_J) & (end_I < end_J))
stilf_relation_during(time8,time7)

# Apply a relation 'contains' (start_I < start_J) & (end_I > end_J)
stilf_relation_contains(time7,time8)

# Apply a relation 'finishes' (start_I > start_J) & (end_I == end_J)
stilf_relation_finishes(time4,time3)

# Apply a relation 'finished by' (start_I < start_J) & (end_I == end_J)
stilf_relation_finished_by(time3,time4)

# Apply a relation 'equals' (start_I == start_J) & (end_I == end_J)
stilf_relation_equals(time3,time4)

# Apply a relation 'in'  (during(first_interval, second_interval) | 
# starts(first_interval, second_interval) | 
# finishes(first_interval, second_interval))
stilf_relation_in(time8,time7)

# Apply a relation 'following' (meets(first_interval, second_interval) |
# before(first_interval, second_interval))
stilf_relation_following(time1,time3)

# Apply a relation 'preceding' (met_by(first_interval, second_interval) |
# after(first_interval, second_interval))
stilf_relation_preceding(time1,time3)
         
           
}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
