% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_load_data.R
\name{example_csv_file}
\alias{example_csv_file}
\title{Example CSV File}
\description{
Dataset of example with input data in format csv to run some functions this package.With data from a particular region from Sinop municipality, Mato Grosso, Brazil.
}
