% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_plot_maps_events.r
\docType{data}
\name{lucC_plot_barplot_events}
\alias{lucC_plot_barplot_events}
\title{Plot Barplot Maps with lucC Events}
\usage{
lucC_plot_barplot_events (data_tb = NULL, 
custom_palette = FALSE, RGB_color = NULL, pixel_resolution = 250)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values longitude and latitude and other values}

\item{custom_palette}{Boolean. A TRUE or FALSE value. If TRUE, user will provide its own color palette setting! Default is FALSE}

\item{RGB_color}{Character. A vector with color names to map legend, for example, c("Green","Blue"). Default is setting scale_colour_hue}

\item{pixel_resolution}{Numeric. Is a spatial resolution of the pixel. Default is 250 meters considering MODIS 250 m. See more at \url{https://modis.gsfc.nasa.gov/about/specifications.php}.}
}
\value{
Plot a barplot in Y axis in square kilometers (Area km^2) = (Number of pixel *(pixel_resolution*pixel_resolution))/(1000*1000)
}
\description{
Plot barplot over time
}
\examples{
\dontrun{

library(lucC)

lucC_starting_point()

# open a JSON file example
file_json = "./inst/extdata/patterns/example_TWDTW.json"

# open file JSON
input_tb_raw_json <- file_json \%>\%
  lucC_fromJSON()
input_tb_raw_json

# plot maps input data
lucC_plot_maps_input(input_tb_raw_json, EPSG_WGS84 = TRUE, 
custom_palette = TRUE, RGB_color = c("#FFB266", "#1b791f", 
"#929e6e", "#f5e7a1"))

# define interval
time_ex1 <- lucC_interval("2002-01-01", "2014-01-01")

# apply predicate occur
ts_occur1 <- lucC_predicate_occur(geo_objects = input_tb_raw_json, 
object_properties = "Forest", event_time_intervals = time_ex1)
ts_occur1

# events over input map
lucC_plot_maps_events(ts_occur1, EPSG_WGS84 = TRUE, 
custom_palette = TRUE, RGB_color = c("#FFB266", "#1b791f", 
"#929e6e", "#f5e7a1"), shape_point = 0, colour_point = "black", 
size_point = 2.3)

lucC_plot_barplot_events(ts_occur1, custom_palette = TRUE, 
RGB_color = "#929e6e", pixel_resolution = 250) 

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
