% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_allen_relations.r
\docType{data}
\name{lucC_relation_equals}
\alias{lucC_relation_equals}
\title{Allen Relation Equals}
\usage{
lucC_relation_equals(first_interval , second_interval)
}
\arguments{
\item{first_interval}{lucC_interval. An interval between two dates.}

\item{second_interval}{lucC_interval. An interval between two dates.}
}
\value{
Logical value if interval are TRUE or FALSE
}
\description{
Provide an Allen's interval relation to classified time series data. And return a logical value if an interval is TRUE or FALSE
}
\examples{
\dontrun{

library(lucC)

lucC_starting_point()

# create some examples of intervals
time3 <- lucC_interval("2011-10-01","2011-11-01")
time4 <- lucC_interval("2011-10-01","2011-11-01")

# Apply a relation 'equals' (start_I == start_J) & (end_I == end_J)
lucC_relation_equals(time3,time4)
           
}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
