% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_allen_relations_extra.r
\docType{data}
\name{lucC_relation_follows}
\alias{lucC_relation_follows}
\title{Allen Relation Follows}
\usage{
lucC_relation_follows(first_interval , second_interval)
}
\arguments{
\item{first_interval}{lucC_interval. An interval between two dates.}

\item{second_interval}{lucC_interval. An interval between two dates.}
}
\value{
Logical value if interval are TRUE or FALSE
}
\description{
Provide an extra Allen's interval relation to classified time series data. And return a logical value if an interval is TRUE or FALSE
}
\examples{
\dontrun{

library(lucC)

lucC_starting_point()

# create some examples of intervals
time1 <- lucC_interval("2011-09-01","2011-10-01")
time3 <- lucC_interval("2011-10-01","2011-11-01")

# Apply a relation 'follows' (meets(first_interval, second_interval) |
# before(first_interval, second_interval))
lucC_relation_follows(time1,time3)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
