% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_allen_relations_extra.r
\docType{data}
\name{lucC_relation_in}
\alias{lucC_relation_in}
\title{Allen Relation In}
\usage{
lucC_relation_in(first_interval , second_interval)
}
\arguments{
\item{first_interval}{lucC_interval. An interval between two dates.}

\item{second_interval}{lucC_interval. An interval between two dates.}
}
\value{
Logical value if interval are TRUE or FALSE
}
\description{
Provide an extra Allen's interval relation to classified time series data. And return a logical value if an interval is TRUE or FALSE
}
\examples{
\dontrun{

library(lucC)

lucC_starting_point()

# create some examples of intervals
time7 <- lucC_interval("2011-08-01","2011-09-15")
time8 <- lucC_interval("2011-08-15","2011-08-29")

# Two interval I and J, with:
# I <- ("2011-08-15","2011-08-29")
# J <- ("2011-09-01","2011-10-01")
# end_I equals to the final value of the interval, I == "2011-08-29"
# start_I equals to the begin value of the interval, I == "2011-08-15"
# end_J equals to the final value of the interval, J == "2011-10-01"
# start_J equals to the begin value of the interval, J == "2011-09-01"

# Apply a relation 'in'  (during(first_interval, second_interval) | 
# starts(first_interval, second_interval) | 
# finishes(first_interval, second_interval))
lucC_relation_in(time8,time7)

           
}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
