% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_applyTWDTW.R
\docType{data}
\name{lucC_TWDTW_fromSITS}
\alias{lucC_TWDTW_fromSITS}
\title{Convert result of sits_TWDTW tibble to lucC format}
\usage{
lucC_TWDTW_fromSITS (data_tb = NULL)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values in sits format}
}
\value{
Tibble with set of points classified by TWDTW method
}
\description{
Convert a tibble classified using sits_TWDTW 
to lucC format (longitude, latitude, start_date, end_date, label, id, index)
}
\examples{
\dontrun{

library(lucC)
library(sits)

lucC_starting_point()

# open data with time series set
data("example_ts")
example_ts

# read a pattern table from a JSON file
patterns.tb <- sits_getdata("./inst/extdata/patterns/example_temporal_pattern.json")
patterns.tb

# plot patterns
sits_plot(patterns.tb, type = "patterns")

# only this bands have in patterns
bands <- c("ndvi", "evi", "nir")

# classification using sits_TWDTW (using sits package)
example_sits_TWDTW <- sits_TWDTW(example_ts, patterns.tb, bands)
example_sits_TWDTW

# convert example_sits_TWDTW from sits package to lucC format
example_convert_sits_TWDTW <- lucC_TWDTW_fromSITS(data_tb = example_sits_TWDTW)
example_convert_sits_TWDTW

# plot maps
lucC_plot_maps_input(example_convert_sits_TWDTW, EPSG_WGS84 = TRUE, 
custom_palette = TRUE, RGB_color = c("#FFB266", "#1b791f", "#929e6e", "#f5e7a1")) 


}

}
\seealso{
sits details at https://github.com/gilbertocamara/sits/
package
}
\author{
Adeline M. Maciel
}
\keyword{datasets}
