% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_applyTWDTW.R
\docType{data}
\name{stilf_applyTWDTW}
\alias{stilf_applyTWDTW}
\title{Stilf Apply TWDTW from SITS Format}
\usage{
stilf_applyTWDTW (data_tb = NULL, 
patterns_tb = NULL, bands = NULL)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values in sits format}

\item{patterns_tb}{Tibble. Temporal patterns in sits format. See more in sits 
package}

\item{bands}{Character. Set of variables with values of spectral bands}
}
\value{
Tibble with set of points  classified by TWDTW method
}
\description{
Classify a set of time series in sits format to stilf format
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

library(sits)

zip_point_tb <- unzip("./inst/list_time_series_Area1_Sinop.zip")
point_tb <- sits::sits_getdata(zip_point_tb)

ts_data <- point_tb

# read a pattern table from a JSON file
patterns_tb <- sits_getdata("./inst/patterns/patterns_
Damien_Ieda_Rodrigo_14classes_3bands_Rename_Labels_Sep.json")

# only this bands have in patterns
bands <- c("ndvi", "evi", "nir")

# plot patterns
sits_plot(patterns_tb, type = "patterns")

# json have other attributes
values <- c("Index", "ndvi", "evi", "nir") # 

# remove other attributes
ts_data$time_series <- lapply(ts_data$time_series,function(p){
  p <- p[,values, drop = FALSE]
})

# classify with TWDTW and return a format to stilf
res <- stilf_applyTWDTW(ts_data, patterns_tb, bands)

res


}

}
\seealso{
sits details at https://github.com/gilbertocamara/sits/
package
}
\author{
Adeline M. Maciel
}
\keyword{datasets}
