% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_load_data.R
\name{stilf_data}
\alias{stilf_data}
\title{data File}
\description{
Dataset of example with input data in format RData to run some functions this package. With data from a particular region from Santa Carmem municipality, Mato Grosso, Brazil.
}
