% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_load_data.R
\docType{data}
\name{stilf_fromJSON}
\alias{stilf_fromJSON}
\title{Open JSON file}
\usage{
stilf_fromJSON (path_json_file = NULL)
}
\arguments{
\item{path_json_file}{Character. Name path and file to open JSON file data}
}
\value{
JSON format to file
}
\description{
Open the input data in JSON format in an user's directory
}
\examples{
\dontrun{
# Open a data example
library(stilf)

json_file = "./inst/example_json_file.json"

# open a json data and store it format stilf_input_data
input_tb_json <- json_file \%>\% 
  stilf_fromJSON()  
input_tb_json


}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
