% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_interval.R
\docType{data}
\name{stilf_interval}
\alias{stilf_interval}
\title{STILF Intervals}
\usage{
stilf_interval(first_date , second_date)
}
\arguments{
\item{first_date}{Date. An date value like '2012-03-02'.}

\item{second_date}{Date. An date value like '2013-03-02'.}
}
\value{
Interval value with two dates
}
\description{
Provide a stilf_interval, an input pattern format used by Allen's relations. And return a value of interval
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# create some examples of intervals
time1 <- stilf_interval("2011-09-01","2011-10-01")
time2 <- stilf_interval("2011-09-15","2011-11-01")


}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
