% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_plot_maps_events.r
\docType{data}
\name{stilf_plot_barplot_events}
\alias{stilf_plot_barplot_events}
\title{Plot Barplot Maps with STILF Events}
\usage{
stilf_plot_barplot_events (data_tb = NULL, 
custom_palette = FALSE, RGB_color = NULL, pixel_resolution = 250)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values longitude and latitude and other values}

\item{custom_palette}{Boolean. A TRUE or FALSE value. If TRUE, user will provide its own color palette setting! Default is FALSE}

\item{RGB_color}{Character. A vector with color names to map legend, for example, c("Green","Blue"). Default is setting scale_colour_hue}

\item{pixel_resolution}{Numeric. Is a spatial resolution of the pixel. Default is 250 meters considering MODIS 250 m. See more at \url{https://modis.gsfc.nasa.gov/about/specifications.php}.}
}
\value{
Plot a barplot in Y axis in square kilometers (Area km^2) = (Number of pixel *(pixel_resolution*pixel_resolution))/(1000*1000)
}
\description{
Plot barplot over time
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# open a JSON file example
file_json = "./inst/example_json_Sinop_part.json"

# open file JSON
input_tb_raw_json <- file_json \%>\% 
  stilf_fromJSON() 
input_tb_raw_json

# plot maps input data
stilf_plot_maps_input(input_tb_raw_json, EPSG_WGS84 = TRUE)

# define interval
time_ex1 <- stilf_interval("2002-01-01", "2014-01-01")
 
# apply predicate occur
ts_occur1 <- stilf_predicate_occur(geo_objects = input_tb_raw_json, 
object_properties = "Forest", event_time_intervals = time_ex1)

# events over input map
stilf_plot_maps_events(ts_occur1, EPSG_WGS84 = TRUE, 
custom_palette = FALSE, size_square = 1)

# plot barplot of events
stilf_plot_barplot_events(ts_occur1, custom_palette = FALSE)


}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
