% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_plot_maps_events.r
\docType{data}
\name{stilf_plot_maps_events}
\alias{stilf_plot_maps_events}
\title{Plot Events over Maps with STILF}
\usage{
stilf_plot_maps_events (data_tb = NULL, EPSG_WGS84 = TRUE, size_square = 1)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values longitude and latitude and other values}

\item{EPSG_WGS84}{Character. A reference coordinate system. If TRUE, the values of latitude and longitude alredy use this coordinate system, if FALSE, the data set need to be transformed}

\item{size_square}{Numeric. Is a size of the square around pixels in plot. Default is 1}
}
\value{
Plot with input data as colored map
}
\description{
Plot map ggplot2 for all events discovered in input data
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# open a JSON file example
file_json = "./inst/example_json_Sinop_part.json"

# open file JSON
input_tb_raw_json <- file_json \%>\% 
  stilf_fromJSON() 
input_tb_raw_json

# plot maps input data
stilf_plot_maps_input(input_tb_raw_json, EPSG_WGS84 = TRUE)

# define interval
time_ex1 <- stilf_interval("2002-01-01", "2014-01-01")
 
# apply predicate occur
ts_occur1 <- stilf_predicate_occur(geo_objects = input_tb_raw_json, 
object_properties = "Forest", event_time_intervals = time_ex1)

# events over input map
stilf_plot_maps_events(ts_occur1, EPSG_WGS84 = TRUE, size_square = 1)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
