% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_plot_maps_input.r
\docType{data}
\name{stilf_plot_maps_input}
\alias{stilf_plot_maps_input}
\title{Plot Input Maps}
\usage{
stilf_plot_maps_input (data_tb = NULL, EPSG_WGS84 = FALSE)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values longitude and latitude and other values}

\item{EPSG_WGS84}{Character. A reference coordinate system. If TRUE, the values of latitude and longitude alredy use this coordinate system, if FALSE, the data set need to be transformed}
}
\value{
Plot with input data as colored map
}
\description{
Plot map ggplot2 for all input data
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# open a CSV file example
file_json = "./inst/example_json_Sinop_part.json"

# open file JSON
input_tb_raw_json <- file_json \%>\% 
  stilf_fromJSON() 
input_tb_raw_json

# plot maps input data
stilf_plot_maps_input(input_tb_raw_json, EPSG_WGS84 = TRUE)


}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
