% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_plot_maps_events.r
\docType{data}
\name{stilf_plot_sequence_events}
\alias{stilf_plot_sequence_events}
\title{Plot Sequence Maps with STILF Events}
\usage{
stilf_plot_sequence_events (data_tb = NULL, show_y_index = TRUE, 
start_date = "2000-01-01", end_date = "2016-12-31")
}
\arguments{
\item{data_tb}{Tibble. A tibble with values longitude and latitude and other values}

\item{show_y_index}{Boolean. TRUE/FALSE to show the index values in the axis y of the graphic}

\item{start_date}{Date. A start date to plot in sequence in format (ymd), '2011-01-01'}

\item{end_date}{Date. A end date to plot in sequence in format (ymd), '2013-01-01'}
}
\value{
Plot sequence time series as lines
}
\description{
Plot time series as sequence of lines over time
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# open a JSON file example
file_json = "./inst/example_json_Sinop_part.json"

# open file JSON
input_tb_raw_json <- file_json \%>\% 
  stilf_fromJSON() 
input_tb_raw_json

# plot maps input data
stilf_plot_maps_input(input_tb_raw_json, EPSG_WGS84 = TRUE)

# define interval
time_ex1 <- stilf_interval("2002-01-01", "2014-01-01")
 
# apply predicate occur
ts_occur1 <- stilf_predicate_occur(geo_objects = input_tb_raw_json, 
object_properties = "Forest", event_time_intervals = time_ex1)

# events over input map
stilf_plot_maps_events(ts_occur1, EPSG_WGS84 = TRUE, size_square = 1)

# plot sequence of events
stilf_plot_sequence_events(ts_occur1, show_y_index = TRUE, 
start_date = "2000-01-01",  end_date = "2016-12-31")

  
}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
