% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_predicates_holds.r
\docType{data}
\name{stilf_predicate_holds}
\alias{stilf_predicate_holds}
\title{Predicate Allen Holds}
\usage{
stilf_predicate_holds (locations = NULL, location_properties = NULL, 
time_intervals = stilf_interval("2000-01-01", "2004-01-01"))
}
\arguments{
\item{locations}{Tibble. A tibble with values longitude and latitude and other values}

\item{location_properties}{Character. Name of value present in a row of the tibble, such as 'Forest' or other value}

\item{time_intervals}{Interval. A interval of time to verify if location_properties is over or 
not in stilf_interval format. Given a tibble with values, will be asserts if that location_properties 
of locations holds during a time interval.}
}
\value{
Tibble with all events hold during a time interval
}
\description{
Provide a predicate of Allen's which asserts that a property holds 
during a time interval. Return a tibble with value within defined interval
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

json_file = "./inst/extdata/patterns/example_TWDTW.json"

input_tb_json <- json_file \%>\% 
  stilf_fromJSON()  
input_tb_json

# example of application
time_ex1 <- stilf_interval("2001-01-01", "2003-01-01")
time_ex1
time_ex2 <- stilf_interval("2005-01-01", "2010-01-01")
time_ex2

# location_properties
properties <- "Forest"

# example predicate holds
stilf_predicate_holds(locations = input_tb_json, 
location_properties = "Forest", time_intervals = time_ex1)

stilf_predicate_holds(locations = input_tb_json, 
location_properties = properties, time_intervals = time_ex2)

}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
