% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_predicates_occurs_and_holds.r
\docType{data}
\name{stilf_predicate_occur}
\alias{stilf_predicate_occur}
\title{Predicate Allen Occur}
\usage{
stilf_predicate_occur (geo_objects = NULL, object_properties = NULL, 
event_time_intervals = stilf_interval("2000-01-01", "2004-01-01"))
}
\arguments{
\item{geo_objects}{Tibble. A tibble with values longitude and latitude and other values}

\item{object_properties}{Character. Name of value present in a row of the tibble, such as 'Forest' or other value}

\item{event_time_intervals}{Interval. A interval of time to verify if event about object_properties occurs or not in stilf_interval format}
}
\value{
Tibble with all events happened over a time interval
}
\description{
Provide a predicate of Allen's which asserts that an event happened over a time interval
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

file_json = "./inst/extdata/patterns/example_TWDTW.json"
input_tb_raw_json <- file_json \%>\% 
  stilf_fromJSON() 
input_tb_raw_json

# plot maps input data
stilf_plot_maps_input(input_tb_raw_json, EPSG_WGS84 = TRUE)

# define interval
time_ex1 <- stilf_interval("2002-01-01", "2014-01-01")
time_ex1

# using occur
ts_occur1 <- stilf_predicate_occur(geo_objects = input_tb_raw_json, 
object_properties = "Pasture", event_time_intervals = time_ex1)
ts_occur1

ts_occur2 <- stilf_predicate_occur(geo_objects = input_tb_raw_json, 
object_properties = "Forest", event_time_intervals = time_ex1)
ts_occur2

# events over input map
stilf_plot_maps_events(ts_occur1, EPSG_WGS84 = TRUE)
stilf_plot_maps_events(ts_occur2, EPSG_WGS84 = TRUE)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
