% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_allen_relations.r
\docType{data}
\name{stilf_relation_after}
\alias{stilf_relation_after}
\title{Allen Relation After}
\usage{
stilf_relation_after(first_interval , second_interval)
}
\arguments{
\item{first_interval}{stilf_interval. An interval between two dates.}

\item{second_interval}{stilf_interval. An interval between two dates.}
}
\value{
Logical value if interval are TRUE or FALSE
}
\description{
Provide an Allen's interval relation to classified time series data. And return a logical value if an interval is TRUE or FALSE
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# create some examples of intervals
time1 <- stilf_interval("2011-09-01","2011-10-01")
time4 <- stilf_interval("2011-10-01","2011-11-01")

# Apply a relation 'after' (start_I > end_J)
stilf_relation_after(time4,time1)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
