% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_allen_relations_extra.r
\docType{data}
\name{stilf_relation_following}
\alias{stilf_relation_following}
\title{Allen Relation Following}
\usage{
stilf_relation_following(first_interval , second_interval)
}
\arguments{
\item{first_interval}{stilf_interval. An interval between two dates.}

\item{second_interval}{stilf_interval. An interval between two dates.}
}
\value{
Logical value if interval are TRUE or FALSE
}
\description{
Provide an extra Allen's interval relation to classified time series data. And return a logical value if an interval is TRUE or FALSE
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# create some examples of intervals
time1 <- stilf_interval("2011-09-01","2011-10-01")
time3 <- stilf_interval("2011-10-01","2011-11-01")

# Apply a relation 'following' (meets(first_interval, second_interval) |
# before(first_interval, second_interval))
stilf_relation_following(time1,time3)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
