% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_allen_relations.r
\docType{data}
\name{stilf_relation_met_by}
\alias{stilf_relation_met_by}
\title{Allen Relation Met By}
\usage{
stilf_relation_met_by(first_interval , second_interval)
}
\arguments{
\item{first_interval}{stilf_interval. An interval between two dates.}

\item{second_interval}{stilf_interval. An interval between two dates.}
}
\value{
Logical value if interval are TRUE or FALSE
}
\description{
Provide an Allen's interval relation to classified time series data. And return a logical value if an interval is TRUE or FALSE
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# create some examples of intervals
time1 <- stilf_interval("2011-09-01","2011-10-01")
time5 <- stilf_interval("2011-08-01","2011-09-01")

# Apply a relation 'met by' (end_J == start_I)
stilf_relation_met_by(time1,time5)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
