% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_standard_date_events.R
\docType{data}
\name{stilf_standard_date_events}
\alias{stilf_standard_date_events}
\title{Stilf Apply TWDTW from SITS Format}
\usage{
stilf_standard_date_events(data_tb = NULL, 
month_year = "09", day_month = "01")
}
\arguments{
\item{data_tb}{Tibble. A tibble with values in sits format}

\item{month_year}{Character. A month of the year in two digits (01 to 12)}

\item{day_month}{Character. A day of the month in two digits (01 to 31). 
Depends the character calendar month.}
}
\value{
Tibble containing two new columns with start_date and end_date in a 
predefined standard range ensurer
}
\description{
Create two new columns in tibble with standard date in 
order to extract events
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

file_json_zip = "./inst/areaSinop/classification_14patterns_3bands_Area1_Sinop.zip"

input_tb_area_Sinop <- file_json_zip \%>\% 
  unzip() \%>\% 
  stilf_fromJSON() 

input_tb_area_Sinop

#remove columns and pass to stilf format
new_area_Sinop <- input_tb_area_Sinop \%>\% 
  select(longitude,latitude,label,start_date,end_date,id,index) \%>\% 
  stilf_data_preparation()

new_area_Sinop

stilf_standard_date_events(data_tb = df, month_year = "09", 
day_month = "01")

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
