% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_preparing_data.R
\docType{data}
\name{stilf_starting_point}
\alias{stilf_starting_point}
\title{Format data as tibble format}
\usage{
stilf_starting_point()
}
\value{
Change the global default setting
}
\description{
Change the global default setting to not auto-convert to factors
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
