% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_save_raster_GeoTIFF.R
\docType{data}
\name{stilf_toGeoTIFF}
\alias{stilf_toGeoTIFF}
\title{Save tibble as Raster in Folder}
\usage{
stilf_toGeoTIFF (data_tb = NULL, path_raster_folder = NULL)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values longitude and latitude and other values}

\item{path_raster_folder}{Character. Name a path folder to save raster images data}
}
\value{
Images in geotiff format to open using SIG
}
\description{
Plot map ggplot2 for all events discovered in input data
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# open a JSON file example
file_json = "./inst/example_json_Sinop_part.json"

input_tb_raw_json <- file_json \%>\% 
  stilf_fromJSON() 
input_tb_raw_json

# save rasters in folder
stilf_toGeoTIFF (input_tb_raw_json, "~/Desktop/raster")

 
}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
