% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_load_data.R
\docType{data}
\name{stilf_toJSON}
\alias{stilf_toJSON}
\title{Save result as JSON format}
\usage{
stilf_toJSON (data_tb, path_json_file = NULL)
}
\arguments{
\item{data_tb}{Tibble. A data frame with input values}

\item{path_json_file}{Character. Name path and file to save JSON file data}
}
\value{
JSON format to file stored
}
\description{
Save the output data in JSON format in an user's directory
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# Example to save data in JSON format
file = "./inst/example_csv_file.csv"

input_tb_csv <- file \%>\% 
  stilf_fromCSV(separator = ",", header_file = TRUE) 
input_tb_csv

output_file = "~/Desktop/output_file.json"
stilf_toJSON(input_tb_csv, output_file)


}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
