% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_applyTWDTW.R
\docType{data}
\name{stilf_TWDTW}
\alias{stilf_TWDTW}
\title{Apply TWDTW Classifier on Time Series Tibble (using the sits package)}
\usage{
stilf_TWDTW (data_tb = NULL, 
patterns_tb = NULL, bands = NULL)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values from sits format}

\item{patterns_tb}{Tibble. Temporal patterns in sits format. See more in sits package}

\item{bands}{Character. Set of variables with values of spectral bands}
}
\value{
Tibble with set of points  classified by TWDTW method
}
\description{
Classify a set of time series from sits_table, using 
sits_getdata directly in the format stilf (longitude, latitude, start_date, end_date, label, id, index)
}
\examples{
\dontrun{

library(stilf)
library(sits)

stilf_starting_point()

# open data with time series set
data("example_ts")
example_ts

# read a pattern table from a JSON file
patterns.tb <- sits_getdata("./inst/extdata/patterns/example_temporal_pattern.json")
patterns.tb

# plot patterns
sits_plot(patterns.tb, type = "patterns")

# only this bands have in patterns
bands <- c("ndvi", "evi", "nir")

# classification using stilf_TWDTW with default parameters from sits
example_stilf_TWDTW <- stilf_TWDTW(example_ts, patterns.tb, bands)
example_stilf_TWDTW

# plot maps
stilf_plot_maps_input(example_stilf_TWDTW, EPSG_WGS84 = TRUE, 
custom_palette = TRUE, RGB_color = c("#FFB266", "#1b791f", "#929e6e", "#f5e7a1")) 

}

}
\seealso{
sits details at https://github.com/gilbertocamara/sits/ package
}
\author{
Adeline M. Maciel
}
\keyword{datasets}
