% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_event_transitions.R
\docType{data}
\name{stilf_event_transitions}
\alias{stilf_event_transitions}
\title{STILF Event Transitions}
\usage{
stilf_event_transitions(data_tb = NULL, properties = NULL, 
time_intervals = stilf_interval("2000-01-01", "2016-12-31"))
}
\arguments{
\item{data_tb}{Tibble. A data frame with input values.}

\item{properties}{Character. A vector with at least 2 strings,
and maximum 10 combination of transitions, 
for example, c("Forest","Cropping").}

\item{time_intervals}{Interval. A interval of time to verify if 
properties is into of stilf_interval format.}
}
\value{
Tibble with event transition
}
\description{
Provide a stilf_event_transitions, a set of event transitions. 
And return a tibble with events discovered
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

data("example_TWDTW")
example_TWDTW

example_3.tb <- example_TWDTW \%>\% 
  stilf_standard_date_events(data_tb = ., month_year = "09", day_month = "01")

example_3.tb

# p = properties of objects :
p1 <- c("Forest", "Pasture", "Single_cropping", "Double_cropping")

# t = interval:
t1 <- stilf_interval("2000-08-01","2017-03-01")

tb <- example_3.tb
output.tb3 <- tb[FALSE,]
coord <- unique(tb$index)

# Apply for each time series based on index
for(x in 1:length(coord)){
  temp.tb <- tb[which(as.character(tb$index) == coord[x]),]
  temp_final.tb <- stilf_event_transitions(temp.tb, properties = p1, 
  time_intervals = t1)
  output.tb3 <- dplyr::bind_rows(output.tb3, temp_final.tb)
}
output.tb3

# plots
stilf_plot_maps_input(example_3.tb, EPSG_WGS84 = TRUE, custom_palette = TRUE, 
RGB_color = c("#FFB266", "#1b791f", "#929e6e", "#f5e7a1"))

stilf_plot_maps_events(output.tb3, EPSG_WGS84 = TRUE, custom_palette = TRUE, 
RGB_color = c("#FFB266", "#1b791f", "#929e6e", "#f5e7a1"), shape_point = 0, 
colour_point = "blue", size_point = 2.3) 

stilf_plot_barplot_events(output.tb3, custom_palette = TRUE, 
RGB_color = c("#FFB266", "#1b791f", "#929e6e", "#f5e7a1"), 
pixel_resolution = 250) 

stilf_plot_sequence_events(output.tb3, show_y_index = TRUE, 
end_date = "2017-03-01", custom_palette = TRUE, 
RGB_color = c("#FFB266", "#1b791f", "#929e6e", "#f5e7a1")) 


}
}
\author{
Adeline M. Maciel
}
\keyword{datasets}
