% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_event_transitions.R
\docType{data}
\name{stilf_event_transitions}
\alias{stilf_event_transitions}
\title{STILF Event Transitions}
\usage{
stilf_event_transitions(data_tb = NULL, properties = NULL, 
time_intervals = stilf_interval("2000-01-01", "2016-12-31"))
}
\arguments{
\item{data_tb}{Tibble. A data frame with input values.}

\item{properties}{Character. A vector with at least 2 strings,
and maximum 10 combination of transitions, 
for example, c("Forest","Cropping").}

\item{time_intervals}{Interval. A interval of time to verify if 
properties is over or not in stilf_interval format.}
}
\value{
Tibble with event transition
}
\description{
Provide a stilf_event_transitions, a set of event transitions. 
And return a tibble with events discovered
}
\author{
Adeline M. Maciel
}
\keyword{datasets}
