% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_event_transitions.R
\docType{data}
\name{stilf_event_transitions}
\alias{stilf_event_transitions}
\title{STILF Event Transitions}
\usage{
stilf_event_transitions(data_tb , properties)
}
\arguments{
\item{data_tb}{Tibble. A data frame with input values.}

\item{properties}{Character. A vector with ate least two strings, 
for example, c("Forest","Cropping").}
}
\value{
Tibble with event transition
}
\description{
Provide a stilf_event_transitions, a set of event transitions. 
And return a tibble with events discovered
}
\author{
Adeline M. Maciel
}
\keyword{datasets}
