% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_load_data.R
\docType{data}
\name{stilf_fromCSV}
\alias{stilf_fromCSV}
\title{Open CSV file}
\usage{
stilf_fromCSV (path_csv_file = NULL, separator = ",", header_file = TRUE)
}
\arguments{
\item{path_csv_file}{Character. Name path and file to open CSV file data}

\item{separator}{Character. Separator to csv, i.e. ',' or ';' ...}

\item{header_file}{Character. Header file csv, if TRUE or FALSE, but in this application is necessary have}
}
\value{
JSON format to file
}
\description{
Open the input data in CSV format in an user's directory
}
\examples{
\dontrun{
# Open a data example
library(stilf)

file_csv = "./inst/example_csv_file.csv"

# read a csv into stilf_input_data format
input_tb_csv <- file_csv \%>\% 
  stilf_fromCSV(separator = ",", header_file = TRUE) 
input_tb_csv

}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
