% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_plot_maps_events.r
\docType{data}
\name{stilf_plot_barplot_events}
\alias{stilf_plot_barplot_events}
\title{Plot Barplot Maps with STILF Events}
\usage{
stilf_plot_barplot_events (data_tb = NULL)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values longitude and latitude and other values}
}
\value{
Plot a barplot in Y axis in (Area km^2) = (Freq*(250*250))/(1000*1000)
}
\description{
Plot barplot over time
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# open file JSON
input_tb_raw_json <- file_json \%>\% 
  stilf_fromJSON() 
input_tb_raw_json

# plot maps input data
stilf_plot_maps_input(input_tb_raw_json, EPSG_WGS84 = TRUE)

# define interval
time_ex1 <- stilf_interval("2002-01-01", "2014-01-01")
 
# apply predicate occur
ts_occur1 <- stilf_predicate_occur(geo_objects = input_tb_raw_json, 
object_properties = "Forest", event_time_intervals = time_ex1)

# events over input map
stilf_plot_maps_events(ts_occur1, EPSG_WGS84 = TRUE, size_square = 1)

# plot barplot of events
stilf_plot_barplot_events(ts_occur1)


}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
