% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_plot_maps_input.r
\docType{data}
\name{stilf_plot_maps_input}
\alias{stilf_plot_maps_input}
\title{Plot Input Maps}
\usage{
stilf_plot_maps_input (data_tb = NULL, EPSG_WGS84 = TRUE, 
custom_palette = FALSE, RGB_color = NULL)
}
\arguments{
\item{data_tb}{Tibble. A tibble with values longitude and latitude and other values}

\item{EPSG_WGS84}{Character. A reference coordinate system. If TRUE, the values of latitude and longitude alredy use this coordinate system, if FALSE, the data set need to be transformed}

\item{custom_palette}{Boolean. A TRUE or FALSE value. If TRUE, user will provide its own color palette setting! Default is FALSE}

\item{RGB_color}{Character. A vector with color names to map legend, for example, c("Green","Blue"). Default is setting scale_colour_hue}
}
\value{
Plot with input data as colored map
}
\description{
Plot map ggplot2 for all input data
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# open a JSON file example
file_json = "./inst/extdata/patterns/example_TWDTW.json"

# open file JSON
input_tb_raw_json <- file_json \%>\%
  stilf_fromJSON()
input_tb_raw_json

# plot maps input data
stilf_plot_maps_input(input_tb_raw_json, EPSG_WGS84 = TRUE, 
custom_palette = TRUE, RGB_color = c("#FFB266", "#1b791f", 
"#929e6e", "#f5e7a1"))


}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
