% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stilf_predicates_occurs_and_holds.r
\docType{data}
\name{stilf_predicate_occur}
\alias{stilf_predicate_occur}
\title{Predicate Allen Occur}
\usage{
stilf_predicate_occur (geo_objects = NULL, object_properties = NULL, 
event_time_intervals = stilf_interval("2000-01-01", "2004-01-01"))
}
\arguments{
\item{geo_objects}{Tibble. A tibble with values longitude and latitude and other values}

\item{object_properties}{Character. Name of value present in a row of the tibble, such as 'Forest' or other value}

\item{event_time_intervals}{Interval. A interval of time to verify if event about object_properties occurs or not in stilf_interval format}
}
\value{
Tibble with all events happened over a time interval
}
\description{
Provide a predicate of Allen's which asserts that an event happened over a time interval
}
\examples{
\dontrun{

library(stilf)

stilf_starting_point()

# open json file
json_file = "./inst/example_json_file.json"

input_tb <- json_file \%>\% 
  stilf_fromJSON()

input_tb

# example of application
time_ex1 <- stilf_interval("2001-01-01", "2003-01-01")
time_ex1
time_ex2 <- stilf_interval("2005-01-01", "2010-01-01")
time_ex2

# object_properties
properties <- "Soybean_Fallow"

# example predicate occur
stilf_predicate_occur(geo_objects = new_a, 
object_properties = "Forest", event_time_intervals = time_ex1)

stilf_predicate_occur(geo_objects = input_tb, 
object_properties = properties, event_time_intervals = time_ex2)

}


}
\author{
Adeline M. Maciel
}
\keyword{datasets}
